[unit converter] - rename directory
